requirements.txt
amqp==5.1.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3
botocore
bs4==0.0.1
celery[redis]
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Flask==2.1.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
pycurl
pyparsing==3.0.9
pytest
python-dotenv==0.21.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
rq==1.11.0
setuptools<58.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.10
user-agent==0.1.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1